backend/requirements.txt
# Em: requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
fastapi-cors
passlib[bcrypt]
python-jose
packaging